% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accele_binary.R
\name{accele_binary}
\alias{accele_binary}
\title{accele_binary}
\usage{
accele_binary(sub_vector, ...)
}
\arguments{
\item{sub_vector}{containing euclidean_norm values of accelerometer data}

\item{...}{}
}
\value{
Returns binary values +1/-1 depending on whether the acceleration in the
sub_vector increased or decreased.
}
\description{
Should primarily be used inside the accele_windslide function after using accele_euclidean
}
\examples{

# Should primarily be used inside the accele_windslide function after using accele_euclidean
df_window <- accele_windslide(df_eucl, FUN = accele_binary, window = 1000)


}
