% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accele_plot.R
\name{accele_plot}
\alias{accele_plot}
\title{accele_plot}
\usage{
accele_plot(x, thresh_or_bi, ...)
}
\arguments{
\item{x}{Data frame made using accele_windslide. Should contain time and
column "accele_threshold" or "accele_binary"}

\item{thresh_or_bi}{Provide either "threshold" or "binary"}
}
\value{
Returns plot with the acceleration vector and values calculated through
either accele_binary or accele_threshold. Use after accele_windowslide.
}
\description{
For plotting data after using accele_euclidean followed by accele_windowslide.
}
\examples{
# If the accele_threshold function was used in accele_windowslide
accele_plot(df_window, thresh_or_bi = "threshold")

# If the accele_binary function was used in accele_windowslide
accele_plot(df_window, thresh_or_bi = "binary")
}
