% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accele_threshold.R
\name{accele_threshold}
\alias{accele_threshold}
\title{accele_threshold}
\usage{
accele_threshold(sub_vector, threshold = 0.5, ...)
}
\arguments{
\item{sub_vector}{Containing euclidean_norm values of accelerometer data}

\item{threshold}{Set threshold value}
}
\value{
Returns the change in euclidean norm value only if the increase or decrease in
acceleration exceeds a threshold value in the window (sub-vector).
}
\description{
Should primarily be used inside the accele_windslide function after using accele_euclidean
}
\examples{
# Should primarily be used inside the accele_windslide function after using accele_euclidean
df_window <- accele_windslide(df_eucl, FUN = accele_threshold, window = 1000, threshold = 0.5)

}
