% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accele_windslide.R
\name{accele_windslide}
\alias{accele_windslide}
\title{accele_windslide}
\usage{
accele_windslide(x, window, FUN, ...)
}
\arguments{
\item{x}{Data frame containing column "euclidean" made using function accel_euclidean}

\item{window}{Which window size to use}

\item{FUN}{Which function to use, accele_binary or accele_threshold}

\item{...}{}
}
\value{
Returns the output of the function provided in FUN in a new column of a data frame
}
\description{
Sliding window function that iteratively works its way through an accelerometer
time series. Uses a data frame containing the acceleration vector of
accelerometer data (created using accele_euclidean) and loops through each
index of the euclidean column in the data frame. Depending on if you use
accele_binary or accele_threshold as FUN, outputs either the binary
+1/-1 dependent on if the acceleration in the window increased or decreased,
or the change in euclidean norm if the acceleration vector change in the
window exceeds a threshold value,
}
\examples{
# If using accele_threshold, you can change the threshold from the defaulty 0.5
df_window <- accele_windslide(df_eucl, FUN = accele_threshold, window = 1000, threshold = 0.5)

# If using accele_binary
df_window <- accele_windslide(df_eucl, FUN = accele_threshold, window = 1000)

}
