% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleslide.R
\docType{package}
\name{acceleslide-package}
\alias{acceleslide-package}
\title{acceleslide: Analyses accelerometer data using a sliding window}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hallgerdur hallgerdur.kolbeinsdottir@ki.se Kolbeinsdottir

}
\keyword{internal}
