% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleslide.R
\docType{package}
\name{acceleslide}
\alias{acceleslide}
\title{acceleslide}
\description{
Package to analyse accelerometer data comtaining time and x,y,z acceleration data.
}
\details{
After loading accelerometer data, you should be able to analyse accelerometer
by running three functions:
\enumerate{
\item accele_euclidean
\item accele_windowslide
\item accele_plot
}

See documentation of each function for more instructions.
}
\author{
\strong{Maintainer}: Hallgerdur hallgerdur.kolbeinsdottir@ki.se Kolbeinsdottir

}
