% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accele_euclidean.R
\name{accele_euclidean}
\alias{accele_euclidean}
\title{accele_euclidean}
\usage{
accele_euclidean(x)
}
\arguments{
\item{x}{Data frame with accelerometer data containing columns x, y and z.}
}
\value{
Calculates Acceleration vector as the euclidean norm of three
acceleration coordinates (x, y and z), and appends to a new column in a data frame.
}
\description{
accele_euclidean
}
\examples{

# Create a new data frame with the column euclidean
df_eucl <- accele_euclidean(example_accel_df)


}
